+++ src/index.ts
@@ -1,6 +1,5 @@
 import contracts from "./contracts/_index";
-import graphql from "./graphql/_index";
-import * as chain from "../typechain/index";
+import * as graphql from "./graphql/_index";
 
 export { contracts, graphql };
 export * from "../typechain/index";